Do not inherit Context off ActiveModel
The `Context` ins't really a model and shouldn't really inherit of
`ActiveModel`
