+++ lib/cloudware/command.rb
@@ -49,7 +49,7 @@ module Cloudware
     end
 
     def context
-      Models::Context.new(region: options.region)
+      Context.new(region: options.region)
     end
     memoize :context
